[REFAC] Refatorada as classes loja, produto e principal, pois precisei ajustar uma tipagem d

+++ GR96002-00000-Diele_Da_Silva_Correa/Principal.java
@@ -54,7 +54,7 @@ public class Principal {
         scanner.nextLine();
 
         System.out.print("Salário base dos funcionarios: ");
-        double slrBase = scanner.nextDouble();
+        int slrBase = scanner.nextInt();
 
         scanner.nextLine();
         System.out.println("\n--- Endereço da loja ---");
